+++ bundles/core/src/main/java/io/wcm/wcm/core/components/impl/models/v1/LayoutContainerV1Impl.java
@@ -34,6 +34,7 @@ import com.adobe.cq.export.json.ComponentExporter;
 import com.adobe.cq.export.json.ExporterConstants;
 import com.adobe.cq.wcm.core.components.models.ContainerItem;
 import com.adobe.cq.wcm.core.components.models.LayoutContainer;
+import com.adobe.cq.wcm.core.components.models.LayoutContainer.LayoutType;
 import com.adobe.cq.wcm.core.components.models.ListItem;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
@@ -69,6 +70,9 @@ public class LayoutContainerV1Impl extends AbstractContainerImpl implements Layo
     return this.delegate.getId();
   }
 
+  /**
+   * @deprecated since 1.13.0-2.22.6- use {@link #getChildren()}
+   */
   @Override
   @JsonIgnore
   @Deprecated(since = "1.13.0-2.22.6")
