+++ bundles/core/src/main/webapp/app-root/components/list/v4/README.md
@@ -1,0 +1,18 @@
+List (v4)
+====
+List component written in HTL that renders a configurable collection of items or content.
+
+Extends the [List AEM Sites Core Component][extends-component]
+
+## Resource Type
+```
+wcm-io/wcm/core/components/list/v4/list
+```
+
+## Enhanced Features
+
+* Inherits all features from its [super component][extends-component]
+* Uses [wcm.io Link Handler][wcmio-handler-link] for processing the list item links
+
+[extends-component]: https://github.com/adobe/aem-core-wcm-components/tree/master/content/src/content/jcr_root/apps/core/wcm/components/list/v4/list
+[wcmio-handler-link]: https://wcm.io/handler/link/
